+++ t/lib/GPForum/Test/ModerationRow.pm
@@ -1,0 +1,26 @@
+package GPForum::Test::ModerationRow;
+
+use strict;
+use warnings;
+
+use Mojo::Base -base;
+
+our $VERSION = '0.001';
+
+has data => sub { return {}; };
+
+sub get_column {
+    my ( $self, $column ) = @_;
+
+    return $self->data->{$column};
+}
+
+sub update {
+    my ( $self, $changes ) = @_;
+
+    $self->data( { %{ $self->data }, %{$changes} } );
+
+    return $self;
+}
+
+1;